Update
Version working nearly as intended, only remains to add the logic for sending all the files in the SD card, may be in the future it would be important to add to the event class a flag which specifies if the Event has previously been tried to send. So that priorities can be assigned.